Refactored reservations to have their own state
